backend/requirements.txt
flask==2.0.3
werkzeug==2.0.3
flask-cors==3.0.10
flask-bcrypt==1.0.1
pymongo==4.3.3
PyJWT==2.6.0
